requirements.txt
folium
exifread
pillow
hashlib
matplotlib
